feat(repl): change flux command to be a repl
BREAKING CHANGE: The Flux CLI API has completely changed
